[DO NOT MERGE] ACPI: scan: prevent double enumeration of MDIO bus children
The MDIO bus is responsible for probing and registering its respective
children, such as PHYs or other kind of devices.

It is required that ACPI scan code should not enumerate such
devices, leaving this task for the generic MDIO bus routines,
which are initiated by the controller driver.

This patch prevents unwanted enumeration of the devices by setting
'enumeration_by_parent' flag, depending on whether their parent
device is a member of a known list of MDIO controllers. For now,
the Marvell MDIO controllers' IDs are added.

TODO: move to a solution based on MDIOSerialBus.

Signed-off-by: Marcin Wojtas <mw@semihalf.com>

tmp